# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mister Moderator API
  description: |-
    API мини-приложения **«Мистер Модератор»** для MAX — помощника старосты студенческой группы:
    напоминания, задания, долги, экзамены, материалы, почта и роли внутри группы.

    ### Авторизация
    Мини-приложение передаёт в каждом запросе заголовок `X-Max-Init-Data` — строку
    `WebApp.initData` из MAX Bridge. Сервер проверяет её HMAC-подпись токеном бота.

    * `STRICT_AUTH=true` — заголовок обязателен и подпись проверяется; иначе ответ `401`.
    * `STRICT_AUTH=false` (демо-режим, по умолчанию) — без заголовка запросы выполняются
      от тестового старосты (`max_user_id = test_user_001`): при `USE_TEST_DATA=true` это
      староста группы `iu7-42b` из `backend/seed/test_data.json`, иначе автоматически создаётся
      «Тест Тестов» в группе `TEST-GROUP-01`. Если заголовок передан, пользователь берётся
      из его поля `user` **без проверки подписи** — так проверяются остальные роли.

    ### Роли
    `starosta` (4) → `zam` (3) → `proforg` (2) → `student` (1). Права каждой роли —
    в ответе `GET /api/roles`. При нехватке прав — `403`.

    ### Формат ошибок
    `{"detail": "<сообщение>"}` или `{"detail": {"error": {"code": "...", "message": "..."}}}`.
    Ошибки валидации тела запроса — `422` в стандартном формате FastAPI.
  version: 1.0.0
paths:
...
  /api/debts:
    get:
      tags:
      - Долги
      summary: Список долгов
      operationId: list_debts_api_debts_get
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Debt'
                title: Response List Debts Api Debts Get
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: 'С правом `debts.view.all` — долги всей группы, иначе только свои. `status`: `active`
        или `overdue` — вычисляется по дедлайну.'
      security:
      - MaxInitData: []
      - {}
    post:
      tags:
      - Долги
      summary: Добавить долг
      operationId: create_debt_api_debts_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDebtRequest'
            example:
              studentId: '2'
              subject: Математический анализ
              type: Экзамен
              deadline: '2026-10-15'
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Debt'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '400':
          description: Некорректные данные (например, неверный формат даты) или пользователь не в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Себе — право `debts.create.own`; любому участнику своей группы — право `debts.edit`.
        `deadline` — `YYYY-MM-DD`. `404`, если участник с `studentId` не найден в группе.
      security:
      - MaxInitData: []
      - {}
  /api/debts/{debt_id}:
    put:
      tags:
      - Долги
      summary: Изменить долг
      operationId: update_debt_api_debts__debt_id__put
      parameters:
      - name: debt_id
        in: path
        required: true
        schema:
          type: integer
          title: Debt Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDebtRequest'
            example:
              subject: Матанализ
              deadline: '2026-10-20'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '400':
          description: Некорректные данные (например, неверный формат даты) или пользователь не в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Частичное обновление. С правом `debts.edit` — любой долг своей группы, без него — только
        свой. Переназначить долг другому участнику (`studentId`) может только обладатель `debts.edit`.
      security:
      - MaxInitData: []
      - {}
    delete:
      tags:
      - Долги
      summary: Удалить (закрыть) долг
      operationId: delete_debt_api_debts__debt_id__delete
      parameters:
      - name: debt_id
        in: path
        required: true
        schema:
          type: integer
          title: Debt Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: С правом `debts.edit` — любой долг своей группы, без него — только свой.
      security:
      - MaxInitData: []
      - {}
...
  /api/mail:
    get:
      tags:
      - Почта
      summary: Входящие письма
      operationId: list_mail_api_mail_get
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MailItem'
                title: Response List Mail Api Mail Get
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Письма группы из подключённых ящиков деканата, кафедр и преподавателей. **MVP:** синхронизация
        с почтовыми серверами не реализована — список берётся из БД (тестовые данные).
      security:
      - MaxInitData: []
      - {}
  /api/mail/{mail_id}/forward:
    post:
      tags:
      - Почта
      summary: Переслать письмо в группу
      operationId: forward_mail_api_mail__mail_id__forward_post
      parameters:
      - name: mail_id
        in: path
        required: true
        schema:
          type: integer
          title: Mail Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardMailResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `mail.forward`. Бот отправляет письмо (отправитель, тема, превью) личным сообщением
        каждому участнику группы. `sentTo` — сколько сообщений доставлено.
      security:
      - MaxInitData: []
      - {}
  /api/mail/refresh:
    post:
      tags:
      - Почта
      summary: Проверить новые письма
      operationId: refresh_mail_api_mail_refresh_post
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshMailResponse'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: '**MVP-заглушка:** синхронизация по IMAP не реализована, `newMessages` всегда 0.'
      security:
      - MaxInitData: []
      - {}
  /api/mail/configure:
    post:
      tags:
      - Почта
      summary: Настроить ящики
      operationId: configure_mailboxes_api_mail_configure_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigureMailboxesRequest'
            example:
              mailboxes:
              - id: '1'
                connected: true
                autoForward: false
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '400':
          description: Некорректные данные (например, неверный формат даты) или пользователь не в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `mail.configure`. Для каждого ящика из `mailboxes` обновляет `connected` и `autoForward`
        (только переданные поля). Ящики других групп и несуществующие `id` пропускаются.
      security:
      - MaxInitData: []
      - {}
  /api/mail/mailboxes:
    get:
      tags:
      - Почта
      summary: Подключённые почтовые ящики
      operationId: list_mailboxes_api_mail_mailboxes_get
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mailbox'
                title: Response List Mailboxes Api Mail Mailboxes Get
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `mail.configure`.
      security:
      - MaxInitData: []
      - {}
    post:
      tags:
      - Почта
      summary: Добавить почтовый ящик
      operationId: add_mailbox_api_mail_mailboxes_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMailboxRequest'
            example:
              email: dean@university.example
              label: Деканат
              autoForward: true
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mailbox'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '400':
          description: Некорректные данные (например, неверный формат даты) или пользователь не в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `mail.configure`. Сохраняет адрес и признак автопересылки.
      security:
      - MaxInitData: []
      - {}
  /api/mail/mailboxes/{mailbox_id}:
    delete:
      tags:
      - Почта
      summary: Удалить почтовый ящик
      operationId: remove_mailbox_api_mail_mailboxes__mailbox_id__delete
      parameters:
      - name: mailbox_id
        in: path
        required: true
        schema:
          type: integer
          title: Mailbox Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `mail.configure`.
      security:
      - MaxInitData: []
      - {}
...
components:
  schemas:
    AddMailboxRequest:
      properties:
        email:
          type: string
          title: Email
        label:
          type: string
          title: Label
        autoForward:
          type: boolean
          title: Autoforward
          default: false
      type: object
      required:
      - email
      - label
      title: AddMailboxRequest
...
    ConfigureMailboxesRequest:
      properties:
        mailboxes:
          items:
            $ref: '#/components/schemas/MailboxSettings'
          type: array
          title: Mailboxes
      type: object
      required:
      - mailboxes
      title: ConfigureMailboxesRequest
    CreateDebtRequest:
      properties:
        studentId:
          type: string
          title: Studentid
        subject:
          type: string
          title: Subject
        type:
          type: string
          title: Type
        deadline:
          type: string
          title: Deadline
      type: object
      required:
      - studentId
      - subject
      - type
      - deadline
      title: CreateDebtRequest
...
    Debt:
      properties:
        id:
          type: string
          title: Id
        studentId:
          type: string
          title: Studentid
        studentName:
          type: string
          title: Studentname
        subject:
          type: string
          title: Subject
        type:
          type: string
          title: Type
        deadline:
          type: string
          format: date-time
          title: Deadline
        status:
          type: string
          title: Status
      type: object
      required:
      - id
      - studentId
      - studentName
      - subject
      - type
      - deadline
      - status
      title: Debt
...
    ForwardMailResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        id:
          type: string
          title: Id
        forwardedTo:
          type: string
          title: Forwardedto
        sentTo:
          type: integer
          title: Sentto
          default: 0
      type: object
      required:
      - id
      - forwardedTo
      title: ForwardMailResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MailItem:
      properties:
        id:
          type: string
          title: Id
        from:
          type: string
          title: From
        subject:
          type: string
          title: Subject
        preview:
          type: string
          title: Preview
        source:
          type: string
          title: Source
        autoForward:
          type: boolean
          title: Autoforward
      type: object
      required:
      - id
      - from
      - subject
      - preview
      - source
      - autoForward
      title: MailItem
    Mailbox:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        label:
          type: string
          title: Label
        connected:
          type: boolean
          title: Connected
        autoForward:
          type: boolean
          title: Autoforward
      type: object
      required:
      - id
      - email
      - label
      - connected
      - autoForward
      title: Mailbox
    MailboxSettings:
      properties:
        id:
          type: string
          title: Id
        connected:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Connected
        autoForward:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Autoforward
      type: object
      required:
      - id
      title: MailboxSettings
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
    RefreshMailResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        newMessages:
          type: integer
          title: Newmessages
      type: object
      required:
      - newMessages
      title: RefreshMailResponse
...
    UpdateDebtRequest:
      properties:
        studentId:
          anyOf:
          - type: string
          - type: 'null'
          title: Studentid
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        deadline:
          anyOf:
          - type: string
          - type: 'null'
          title: Deadline
      type: object
      title: UpdateDebtRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    HTTPError:
      title: HTTPError
      type: object
      required:
      - detail
      properties:
        detail:
          description: 'Текст ошибки или объект {error: {code, message}}'
          oneOf:
          - type: string
          - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: permission_denied
            message:
              type: string
              example: Нет права roles.assign
  securitySchemes:
    MaxInitData:
      type: apiKey
      in: header
      name: X-Max-Init-Data
      description: Строка WebApp.initData из MAX Bridge. Обязательна при STRICT_AUTH=true.
servers:
- url: /
  description: 'Текущий хост (через nginx: https://discreetly-enjoyable-koi.cloudpub.ru/)'
- url: http://localhost:8080
...